Split utils in submodule
